+++ src/main/java/kitchenpos/order/supports/factory/OrderCreateFactory.java
@@ -4,7 +4,7 @@ import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.order.domain.OrderTable;
 import kitchenpos.order.domain.OrderType;
-import kitchenpos.order.eatinorders.domain.vo.OrderLineItems;
+import kitchenpos.order.domain.vo.OrderLineItems;
 
 import java.time.LocalDateTime;
 import java.util.UUID;
